requirements.txt
feedparser==6.0.11
openai==1.68.2
python-dotenv==1.0.1
